backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
opencv-python==4.10.0.84
Pillow==11.0.0
numpy==2.2.1
gunicorn==23.0.0
requests==2.32.3

# Machine Learning dependencies
tensorflow==2.20.0
scikit-learn==1.6.1
matplotlib==3.10.1
